reverse renaming of modules
use from __future__ import to avoid namespace collisions
