+++ can/__init__.py
@@ -17,11 +17,11 @@ class CanError(IOError):
 
 from can.listener import Listener, BufferedReader, RedirectReader
 
-from can.formated_io import Logger, Printer, LogReader
-from can.formated_io import ASCWriter
-from can.formated_io import BLFReader, BLFWriter
-from can.formated_io import CSVWriter
-from can.formated_io import SqliteWriter, SqlReader
+from can.io import Logger, Printer, LogReader
+from can.io import ASCWriter
+from can.io import BLFReader, BLFWriter
+from can.io import CSVWriter
+from can.io import SqliteWriter, SqlReader
 
 from can.util import set_logging_level
