+++ can/interface.py
@@ -11,7 +11,7 @@ BACKENDS = {
     'kvaser':           ('can.interfaces.kvaser', 'KvaserBus'),
     'socketcan_ctypes': ('can.interfaces.socketcan', 'SocketcanCtypes_Bus'),
     'socketcan_native': ('can.interfaces.socketcan', 'SocketcanNative_Bus'),
-    'basic_serial':     ('can.interfaces.basic_serial.serial_can', 'SerialBus'),
+    'serial':           ('can.interfaces.serial.serial_can', 'SerialBus'),
     'pcan':             ('can.interfaces.pcan', 'PcanBus'),
     'usb2can':          ('can.interfaces.usb2can', 'Usb2canBus'),
     'ixxat':            ('can.interfaces.ixxat', 'IXXATBus'),
