+++ can/interfaces/__init__.py
@@ -3,6 +3,6 @@
 Interfaces contain low level implementations that interact with CAN hardware.
 """
 
-VALID_INTERFACES = set(['kvaser', 'basic_serial', 'pcan', 'socketcan_native',
+VALID_INTERFACES = set(['kvaser', 'serial', 'pcan', 'socketcan_native',
                         'socketcan_ctypes', 'socketcan', 'usb2can', 'ixxat',
                         'nican', 'iscan', 'vector', 'remote', 'virtual', 'neovi','slcan'])
