--- can/interfaces/basic_serial/__init__.py
@@ -1,1 +1,0 @@
-from can.interfaces.basic_serial.serial_can import SerialBus as Bus
