+++ can/interfaces/serial/__init__.py
@@ -1,0 +1,1 @@
+from can.interfaces.serial.serial_can import SerialBus as Bus
