+++ can/player.py
@@ -10,7 +10,7 @@ import argparse
 import datetime
 
 import can
-from can.formated_io.player import LogReader, MessageSync
+from can.io.player import LogReader, MessageSync
 
 
 def main():
@@ -26,7 +26,7 @@ def main():
 
     parser.add_argument('-c', '--channel',
                         help='''Most backend interfaces require some sort of channel.
-    For example with the basic_serial interface the channel might be a rfcomm device: "/dev/rfcomm0"
+    For example with the serial interface the channel might be a rfcomm device: "/dev/rfcomm0"
     With the socketcan interfaces valid channel examples include: "can0", "vcan0"''')
 
     parser.add_argument('-i', '--interface', dest="interface",
